backend/requirements-api.txt
# Ramanujan Compression — API stack (c.md). Install SDK first: pip install -e .
-r requirements.txt

# Web & async
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic-settings>=2.1.0

# Database
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
alembic>=1.13.0
greenlet>=3.0.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
# passlib 1.7.x expects bcrypt<4.1 (4.1+ removed __about__)
bcrypt>=4.0.1,<4.1

# Redis & rate limit
redis>=5.0.0
slowapi>=0.1.9

# HTTP client for LLM adapters
httpx>=0.26.0

# Celery
celery[redis]>=5.3.0

# Utils
email-validator>=2.1.0
